vNext (2.4): Walls & Weather
You can now set a website as the backdrop of the page, check out the backdrop section of the control panel to get started.
Additionally, weather has been completely rebuilt from the ground up with a new backend, allowing for automatic location detection, and a fresh design for the weather card including the 'feels like' metric for weather.